[General] - Visual studio 2010 project files - updated to Ogre 1.7
[Engine]
- Waymesh pathfindig improvements

+++ trunk/Core/Source/IceGOCAI.cpp
@@ -151,7 +151,7 @@ void GOCAI::ReceiveObjectMessage(Ogre::SharedPtr<ObjectMsg> msg)
 	}
 	if (msg->mName == "CharacterCollisionReport")
 	{
-		NxU32 collisionFlags = msg->mData.GetFloat("collisionFlags");
+		NxU32 collisionFlags = msg->mData.GetInt("collisionFlags");
 	}
 }
