+++ trunk/Core/Source/IceGOCPlayerInput.cpp
@@ -118,7 +118,7 @@ void GOCPlayerInput::ReceiveObjectMessage(Ogre::SharedPtr<ObjectMsg> msg)
 	}
 	if (msg->mName == "CharacterCollisionReport")
 	{
-		NxU32 collisionFlags = msg->mData.GetFloat("collisionFlags");
+		NxU32 collisionFlags = msg->mData.GetInt("collisionFlags");
 	}
 }
